chore: refactors example android project structure
