+++ android/build.gradle
@@ -11,6 +11,7 @@ buildscript {
     classpath "com.android.tools.build:gradle:7.4.0"
     // noinspection DifferentKotlinGradleVersion
     classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
+    classpath "com.facebook.react:react-native-gradle-plugin"
   }
 }
