+++ example/android/app/build.gradle
@@ -93,9 +93,9 @@ android {
 
     compileSdkVersion rootProject.ext.compileSdkVersion
 
-    namespace "com.chargebeereactnativeexample"
+    namespace "com.chargebee.example"
     defaultConfig {
-        applicationId "com.chargebeereactnativeexample"
+        applicationId "com.chargebee.example"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
