+++ example/android/app/src/main/java/com/chargebee/example/MainApplication.java
@@ -1,6 +1,9 @@
-package com.chargebeereactnativeexample;
+package com.chargebee.example;
 
 import android.app.Application;
+
+import com.chargebee.example.BuildConfig;
+import com.chargebee.example.ReactNativeFlipper;
 import com.facebook.react.PackageList;
 import com.facebook.react.ReactApplication;
 import com.facebook.react.ReactNativeHost;
